requirements.txt
keras==2.6.0
keras-preprocessing==1.1.2
lightgbm==3.3.1
matplotlib==3.5.0
nltk==3.6.5
numpy==1.21.4
pandas==1.3.4
pillow==8.4.0
progressbar==2.5
scikit-learn==1.0.1
scipy==1.7.2
tensorboard==2.6.0